+++ .gitignore
@@ -1,1 +1,1 @@
-./node_modules
+node_modules
