+++ frontend/src/constants/clinical_scales.ts
@@ -1,0 +1,59 @@
+
+export const HYPOCHONDRIASIS = {
+  true: [13,37,61,85,109,133,157,181,205,229],
+  false: [1,25,49,73,97,121,145,169,193,217]
+}
+
+export const DEPRESSION = {
+  true: [2,26,50,74,98,122,146,170,194,218],
+  false: [14,38,62,86,110,134,158,182,206,230]
+}
+
+export const DENIAL = {
+  true: [15,39,63,87,111,135,159,183,207,231],
+  false: [3,27,51,75,99,123,147,171,195,219]
+}
+
+export const INTERPERSONAL_PROBLEMS = {
+  true: [16,40,64,88,112,136,160,184,208,232],
+  false: [4,28,52,76,100,124,148,172,196,220]
+}
+
+export const ALIENATION = {
+  true: [5,29,53,77,101,125,149,173,197,221],
+  false: [17,41,65,89,113,137,161,185,209,233]
+}
+
+export const PERSECUTORY_IDEAS = {
+  false: [6,30,54,78,102,126,150,174,198,222],
+  true: [18,42,66,90,114,138,162,186,210,234]
+}
+
+export const ANXIETY = {
+  true: [7,31,55,79,103,127,151,175,199,223],
+  false: [19,43,67,91,115,139,163,187,211,235]
+}
+
+export const THINKING_DISORDER = {
+  true: [20,44,68,92,116,140,164,188,212,236],
+  false: [8,32,56,80,104,128,152,176,200,224]
+}
+
+export const IMPULSE_EXPRESSION = {
+  true: [9,33,57,81,105,129,153,177,201,225],
+  false: [21,45,69,93,117,141,165,189,213,237]
+}
+
+export const SOCIAL_INTROVERSION = {
+  true: [22,46,70,94,118,142,166,190,214,238],
+  false: [10,34,58,82,106,130,154,178,202,226]
+}
+
+export const SELF_DEPRECIATION = {
+  true: [11,35,59,83,107,131,155,179,203,227],
+  false: [23,47,71,95,119,143,167,191,215,239] 
+}
+
+export const DEVIATON = {
+  true: [12,24,36,48,60,72,84,96,108,120,132,144,156,168,180,192,204,216,228,240]
+}
